fix(go/echo-tinygo): remove default wasm_target
While we updated the default target to `wasm32-wasi-preview1`, current
`wash` builds expect that preview1/preview2 components are WIT-first,
and attempt to run bindgen on them.

`echo-tinygo` is *not* WIT-powered, and uses the legacy smithy
interface, which is fine as we are backwards compatible, *but* raises
a problem when `wash` attempts to run bindgen.

This problem was discovered upstream:
https://github.com/wasmCloud/wasmCloud/issues/1241

This commit returns the `wasm_target` default to
`wasm32-unknown-unknown`, in anticipation that an upstream change to
`wash` is merged which does a more stringent check.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>